replaced old lighting function
